feat: 대기열 토큰 관리 레디스 사용

+++ src/main/kotlin/kr/tareun/concert/interfaces/queue/model/QueueResponse.kt
@@ -5,7 +5,7 @@ import java.util.*
 
 data class QueueResponse(
     val uuid: UUID,
-    val remainingQueue: Long
+    val remainingQueue: Int
 ) {
     companion object {
         fun from(queueTokenResult: QueueTokenResult): QueueResponse {
